requirements.txt
appdirs==1.4.3
cachetools==4.0.0
certifi==2019.11.28
django-extensions==2.2.9
django-haystack @ git+https://github.com/django-haystack/django-haystack@ea3f8c961934e8871b89881d7fab09e038191899
django-summernote==0.8.11.6
django-taggit==1.2.0
elasticsearch==5.5.3
pika==1.1.0
Pillow==7.1.2
protobuf==3.11.2
requests==2.22.0
gunicorn==19.6.0
